move mocked users and activities to service tests
