+++ backend/.gitignore
@@ -1,1 +1,2 @@
 #/copySwagger.sh
+/src/main/java/hw/spring/config
